Fix all errors that we throw during tests that throw no errors.
Makes it easier to identify problems in tests.
