+++ extensions/amp-access-laterpay/0.1/test/test-amp-access-laterpay.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {LaterpayVendor} from '../amp-access-laterpay';
+import {LaterpayVendor} from '../laterpay-impl';
 import {toggleExperiment} from '../../../../src/experiments';
 import * as sinon from 'sinon';
